+++ src/components/form/index.js
@@ -1,9 +1,11 @@
 import Input from './input';
+import Select from './select';
 import { usePlugin, installPlugin } from '@/utils';
 
 const VuePlugin = {
   install(Vue) {
     installPlugin(Vue, Input);
+    installPlugin(Vue, Select);
   },
 };
 
@@ -13,4 +15,5 @@ export default VuePlugin;
 
 export {
     Input,
+    Select,
 };